Continued work on PoC
* added some more demo methods and properties
* still in early stage with alot of questions left

+++ OctoAwesome/OctoAwesome.PoC/RecipeItem.cs
@@ -1,3 +1,3 @@
 ﻿namespace OctoAwesome.PoC;
 
-public readonly record struct RecipeItem(string ItemName, int Count);
+public readonly record struct RecipeItem(string ItemName, decimal Count);
